Fix IndexError runtime error
fixes this error:

  File "C:\code\nerfstudio\nerfstudio\data\dataparsers\nerfstudio_dataparser.py", line 232, in _generate_dataparser_outputs
    distortion_params = torch.stack(distort, dim=0)[idx_tensor]
IndexError: tensors used as indices must be long, byte or bool tensors

+++ nerfstudio/data/dataparsers/nerfstudio_dataparser.py
@@ -229,7 +229,7 @@ class Nerfstudio(DataParser):
                 p2=float(meta["p2"]) if "p2" in meta else 0.0,
             )
         else:
-            distortion_params = torch.stack(distort, dim=0)[idx_tensor]
+            distortion_params = torch.stack(distort, dim=0)[idx_tensor.type(torch.long)]
 
         cameras = Cameras(
             fx=fx,
